Update Raymond Mule-Signoff-Wireless Communication Subsystem(Updated).md
Addressed the comments concerning data by fixing constraint #6. The data rate is specified in the analysis section #2. The schematic that makes it seem that two devices are being plugged into the Jetson Nano via USB port is actually only one device. On the schematic, there is a parenthesis on the component RAKDAP1 that specifies if you are looking at the top of the device or the bottom of the device. I did this since there are inputs on the bottom of the RAKDAP1 and the top of it as well.